% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-generate_counts.R
\name{generate_counts}
\alias{generate_counts}
\title{Simulate counts which are distributed using a zero-inflated negative binomial distribution}
\usage{
generate_counts(
  trajectory,
  num_features = 101,
  sample_mean_count = function() runif(1, 100, 1000),
  sample_dispersion_count = function(mean) map_dbl(mean, ~runif(1, ./10, ./4)),
  dropout_probability_factor = 100,
  dropout_rate = 0.2,
  differentially_expressed_rate = 1
)
}
\arguments{
\item{trajectory}{The dynwrap trajectory}

\item{num_features}{Number of features}

\item{sample_mean_count}{Function used to sample the mean expression}

\item{sample_dispersion_count}{Function to sample the dispersion (size) of the negative biniomal given the expression. Higher dispersion values generate less noise}

\item{dropout_probability_factor}{Factor used to calculate the probabilities of dropouts, relative to expression. Higher values (> 10000) have a lot of dropouts, lower values (< 10) have almost none}

\item{dropout_rate}{Base rate of drop-outs}

\item{differentially_expressed_rate}{Percentage of differentially expressed genes}
}
\description{
Simulate counts which are distributed using a zero-inflated negative binomial distribution
}
