% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-generate-dataset.R
\name{generate_trajectory}
\alias{generate_trajectory}
\alias{generate_dataset}
\title{Generate a toy dataset}
\usage{
generate_trajectory(id = "", model = "linear", num_cells = 99,
  allow_tented_progressions = TRUE)

generate_dataset(id = "", model = "linear", num_cells = 99,
  allow_tented_progressions = TRUE, num_features = 101,
  sample_mean_count = function() runif(1, 100, 1000),
  sample_dispersion_count = function(mean) map_dbl(mean, ~runif(1, ./10,
  ./4)), dropout_probability_factor = 100,
  normalise = dynutils::check_packages("dynnormaliser"),
  add_prior_information = TRUE)
}
\arguments{
\item{id}{An id for the dataset}

\item{model}{A model for generating the milestone network. Must be one of:
\itemize{
  \item{a character vector (e.g. \code{"linear"}),}
  \item{a function (e.g. \code{model_linear}),}
  \item{a data frame (e.g. \code{model_linear()})}
}}

\item{num_cells}{The number of cells in each dataset}

\item{allow_tented_progressions}{Whether or not to be able to generate cells as
part of a divergence.}

\item{num_features}{Number of features}

\item{sample_mean_count}{Function used to sample the mean expression}

\item{sample_dispersion_count}{Function to sample the dispersion (size) of the negative biniomal given the expression. Higher dispersion values generate less noise}

\item{dropout_probability_factor}{Factor used to calculate the probabilities of dropouts, relative to expression. Higher values (> 10000) have a lot of dropouts, lower values (< 10) have almost none}

\item{normalise}{Whether or not to normalise the dataset}

\item{add_prior_information}{Whether to add prior information}
}
\description{
Generate a toy dataset
}
