% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-generate_datasets.R
\name{generate_datasets}
\alias{generate_datasets}
\title{Generate toy datasets with dyntoy}
\usage{
generate_datasets(
  models = c("linear", "cyclic", "bifurcating", "multifurcating", "binary_tree",
    "tree", "converging", "diverging_converging", "diverging_with_loops", "looping",
    "connected", "disconnected"),
  num_cells = 99,
  num_features = 101,
  allow_tented_progressions = TRUE,
  sample_mean_count = function() runif(1, 100, 1000),
  sample_dispersion_count = function(mean) map_dbl(mean, ~runif(1, ./10, ./4)),
  dropout_probability_factor = 100,
  normalise = FALSE,
  add_prior_information = TRUE,
  num_replicates = 3
)
}
\arguments{
\item{models}{A list of network models to use to generate the milestone networks.
\itemize{
  \item{a character vector (e.g. \code{c("linear", "bifurcating")}),}
  \item{a list of functions (e.g. \code{list(model_linear, model_bifurcating)}),}
  \item{a list of data frames (e.g. \code{list(model_linear(), model_bifurcating())})}
}}

\item{num_cells}{The number of cells in each dataset}

\item{num_features}{Number of features}

\item{allow_tented_progressions}{Whether or not to be able to generate cells as
part of a divergence.}

\item{sample_mean_count}{Function used to sample the mean expression}

\item{sample_dispersion_count}{Function to sample the dispersion (size) of the negative biniomal given the expression. Higher dispersion values generate less noise}

\item{dropout_probability_factor}{Factor used to calculate the probabilities of dropouts, relative to expression. Higher values (> 10000) have a lot of dropouts, lower values (< 10) have almost none}

\item{normalise}{Whether or not to normalise the dataset}

\item{add_prior_information}{Whether to add prior information}

\item{num_replicates}{How many replicates of each TI type to generate}
}
\description{
Generate toy datasets with dyntoy
}
