% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-generate_progressions.R
\name{generate_progressions}
\alias{generate_progressions}
\title{Generate cells and their progressions along a milestone network}
\usage{
generate_progressions(milestone_network, num_cells = 99,
  allow_tented_progressions = TRUE)
}
\arguments{
\item{milestone_network}{The milestone network along which to generate cells.}

\item{num_cells}{The number of cells in each dataset}

\item{allow_tented_progressions}{Whether or not to be able to generate cells as
part of a divergence.}
}
\description{
Generate cells and their progressions along a milestone network
}
