% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-generate-datasets.R
\name{generate_toy_datasets}
\alias{generate_toy_datasets}
\title{Generate toy datasets with dyngen}
\usage{
generate_toy_datasets(models = c("simple_linear", "linear", "bifurcating",
  "cycle", "consecutive_bifurcating", "trifurcating", "converging", "BA",
  "bifuracting_loop", "bifurcating_cycle"), num_replicates = 3,
  num_cells = 200, num_genes = 100, expression_randomizer = "modules",
  noise_nbinom_size = 20)
}
\arguments{
\item{models}{The types of trajectory to generate}

\item{num_replicates}{How many replicates of each TI type to generate}

\item{num_cells}{The number of cells in each dataset}

\item{num_genes}{The number of genes in each dataset}

\item{expression_randomizer}{How to randomize the expression}

\item{noise_nbinom_size}{The size parameter of the nbinom distribution}
}
\description{
Generate toy datasets with dyngen
}
