% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-generate_trajectory.R
\name{generate_trajectory}
\alias{generate_trajectory}
\title{Generate a trajectory}
\usage{
generate_trajectory(
  id = "",
  model = c("linear", "cyclic", "bifurcating", "multifurcating", "binary_tree", "tree",
    "converging", "diverging_converging", "diverging_with_loops", "looping", "connected",
    "disconnected"),
  num_cells = 99,
  allow_tented_progressions = TRUE
)
}
\arguments{
\item{id}{An id for the dataset}

\item{model}{A model for generating the milestone network. Must be one of:
\itemize{
  \item{a character vector (e.g. \code{"linear"}),}
  \item{a function (e.g. \code{model_linear}),}
  \item{a data frame (e.g. \code{model_linear()})}
}}

\item{num_cells}{The number of cells in each dataset}

\item{allow_tented_progressions}{Whether or not to be able to generate cells as
part of a divergence.}
}
\description{
Generate a trajectory
}
