% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-networks.R
\docType{data}
\name{network_models}
\alias{network_models}
\title{Milestone network models}
\format{An object of class \code{list} of length 8.}
\usage{
network_models
}
\description{
Milestone network models
}
\keyword{datasets}
