% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_datasets}
\alias{toy_datasets}
\title{Toy datasets generated by \code{\link{generate_toy_datasets}}}
\format{TODO}
\usage{
toy_datasets
}
\description{
Toy datasets generated by \code{\link{generate_toy_datasets}}
}
\keyword{datasets}
