% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_tasks}
\alias{toy_tasks}
\title{Toy tasks generated by \code{\link{generate_toy_datasets}}}
\format{TODO}
\usage{
toy_tasks
}
\description{
Toy tasks generated by \code{\link{generate_toy_datasets}}
}
\keyword{datasets}
