% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocitise.R
\name{velocitise}
\alias{velocitise}
\title{Velocitise a toy dataset}
\usage{
velocitise(toy, num_heuristic_tries = 100,
  wanted_geodesic_quantile = 0.05)
}
\arguments{
\item{num_heuristic_tries}{number of heuristical tries}

\item{wanted_geodesic_quantile}{the desired geodesic distance between a cells current state and future state
is this quantile value of all geodesic distances.}
}
\description{
This will cut the number of cells in half.
}
\details{
Uses a heuristic to try to pairwise match half of the cells with the other half of the cells.
After running the heuristic `num_heuristic_tries` times, the best heuristic will be chosen.
}
