% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-generate-dataset.R
\name{generate_dataset}
\alias{generate_dataset}
\title{Generate a toy dataset}
\usage{
generate_dataset(unique_id, model = "linear", num_cells = 99,
  num_genes = 101, noise_nbinom_size = 20,
  allow_tented_progressions = TRUE,
  normalise = dynutils::check_packages("dynnormaliser"))
}
\arguments{
\item{unique_id}{An id for the dataset}

\item{model}{A model for generating the milestone network. Must be one of:
\itemize{
  \item{a character vector (e.g. \code{"linear"}),}
  \item{a function (e.g. \code{model_linear}),}
  \item{a data frame (e.g. \code{model_linear()})}
}}

\item{num_cells}{The number of cells in each dataset}

\item{num_genes}{The number of genes in each dataset}

\item{noise_nbinom_size}{The size parameter of the nbinom distribution}

\item{allow_tented_progressions}{Whether or not to be able to generate cells as
part of a divergence.}

\item{normalise}{Whether or not to normalise the dataset}
}
\description{
Generate a toy dataset
}
