% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-generate-datasets.R
\name{generate_toy_datasets}
\alias{generate_toy_datasets}
\title{Generate toy datasets with dyngen}
\usage{
generate_toy_datasets(trajectory_types = c("simple_linear", "linear",
  "bifurcating", "cycle", "consecutive_bifurcating", "trifurcating",
  "converging", "BA", "bifuracting_loop", "bifurcating_cycle"),
  num_replicates = 3, num_cells = 99, num_genes = 101,
  expression_randomizer = "shift")
}
\arguments{
\item{trajectory_types}{The types of trajectory to generate}

\item{num_replicates}{How many replicates of each TI type to generate}

\item{num_cells}{The number of cells in each dataset}

\item{num_genes}{The number of genes in each dataset}

\item{expression_randomizer}{How to randomize the expression}
}
\description{
Generate toy datasets with dyngen
}
