% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-generate-datasets.R
\name{generate_toy_datasets}
\alias{generate_toy_datasets}
\title{Generate toy datasets with dyntoy}
\usage{
generate_toy_datasets(models = c("linear", "cyclic", "multifurcating",
  "converging", "multifurcating_with_loops", "disconnected"),
  num_replicates = 3, num_cells = 200, num_genes = 100,
  noise_nbinom_size = 20, allow_tented_progressions = TRUE)
}
\arguments{
\item{models}{The types of trajectory to generate}

\item{num_replicates}{How many replicates of each TI type to generate}

\item{num_cells}{The number of cells in each dataset}

\item{num_genes}{The number of genes in each dataset}

\item{noise_nbinom_size}{The size parameter of the nbinom distribution}

\item{allow_tented_progressions}{Whether or not to be able to generate cells as
part of a divergence.}
}
\description{
Generate toy datasets with dyntoy
}
