% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-networks.R
\name{generate_toy_milestone_network}
\alias{generate_toy_milestone_network}
\title{Generate a toy milestone network}
\usage{
generate_toy_milestone_network(model = c("linear", "cyclic", "multifurcating",
  "converging", "diverging_converging", "multifurcating_with_loops",
  "anything_goes", "disconnected"), ...)
}
\arguments{
\item{model}{Which model to use.}

\item{...}{Extra params for each model. See \code{\link{network_models}}}
}
\description{
Generate a toy milestone network
}
