% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-networks.R
\docType{data}
\name{model_linear}
\alias{model_linear}
\alias{model_cyclic}
\alias{model_bifurcating}
\alias{model_diverging}
\alias{model_converging}
\alias{model_diverging_converging}
\alias{model_diverging_with_loops}
\alias{model_multiple_looping}
\alias{model_connected}
\alias{model_disconnected}
\alias{network_models}
\alias{generate_milestone_network}
\title{Milestone network models for generating toy examples}
\format{An object of class \code{list} of length 10.}
\usage{
model_linear(num_milestones = function() rbinom(1, size = 10, 0.25) + 2)

model_cyclic(num_milestones = function() rbinom(1, size = 10, 0.25) + 3)

model_bifurcating(num_branchpoints = function() sample_discrete_uniform(1, 3,
  6))

model_diverging(num_branchpoints = function() rbinom(1, size = 10, 0.25) + 1,
  max_degree = function() sample_discrete_uniform(1, 3, 6))

model_converging(num_branchpoints = function() rbinom(1, size = 10, 0.25) + 1,
  max_degree = function() sample_discrete_uniform(1, 3, 6))

model_diverging_converging(num_modifications = function() rbinom(1, size = 10,
  0.25) + 1, nodes_per_modification = function() sample_discrete_uniform(1, 3,
  6))

model_diverging_with_loops(num_modifications = function() rbinom(1, size = 10,
  0.25) + 1, nodes_per_modification = function() sample_discrete_uniform(1, 3,
  6))

model_multiple_looping(num_modifications = function() rbinom(1, size = 10,
  0.25) + 1, nodes_per_modification = function() sample_discrete_uniform(1, 3,
  6))

model_connected(num_modifications = function() rbinom(1, size = 10, 0.25) + 1,
  nodes_per_modification = function() sample_discrete_uniform(1, 3, 6))

model_disconnected(num_trajectories = rbinom(1, size = 5, 0.25) + 2, ...)

network_models

generate_milestone_network(model = c("linear", "cyclic", "bifurcating",
  "diverging", "converging", "diverging_converging", "diverging_with_loops",
  "looping", "connected", "disconnected"), ...)
}
\arguments{
\item{num_milestones}{The number of milestones in the trajectory (linear, cyclic)}

\item{num_branchpoints}{The number of branchpoints in the trajectory (bifurcating, diverging, converging)}

\item{max_degree}{The maximum degree of a branch node, must be at least 3 (diverging, converging)}

\item{num_modifications}{Number of modifications made to the original trajectory (diverging_converging, diverging_with_loops, multiple_looping, connected)}

\item{nodes_per_modification}{The number of nodes to use per modification (diverging_converging, diverging_with_loops, multiple_looping, connected)}

\item{num_trajectories}{The number of disconnected trajectories to generate}

\item{...}{Parameters to pass to other models. Can be in the form of `linear = list(num_milestones = function() sample(2:8, 1)` or just `num_milestones = 10`.}

\item{model}{Which model to use.}
}
\description{
Milestone network models for generating toy examples

Generate a toy milestone network
}
\keyword{datasets}
