% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_datasets}
\alias{toy_datasets}
\title{Toy datasets generated by \code{\link{generate_datasets}}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 60 rows and 19 columns.}
\usage{
toy_datasets
}
\description{
Toy datasets generated by \code{\link{generate_datasets}}
}
\keyword{datasets}
