% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_tasks}
\alias{toy_tasks}
\title{Toy tasks generated by \code{\link{generate_toy_datasets}}}
\format{TODO}
\usage{
toy_tasks
}
\description{
\code{
  toy_tasks <- dyntoy::generate_toy_datasets(num_replicates = 2); devtools::use_data(toy_tasks, overwrite=TRUE)
}
}
\keyword{datasets}
